worker/requirements.txt

fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
spacy==3.7.4
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
psycopg2-binary==2.9.9
